+++ app/services/api/index.ts
@@ -9,4 +9,5 @@ export * from './schedule';
 export * from './admin';
 export * from './info';
 export * from './uploads';
+export * from './memes';
 export * from './notifications';